refactoring(searchBarAutocomplete): extract useSelectOption
